requirements.txt
nodriver
selenium_driverless
openai
browserforge
camoufox
python-dotenv
pythonbible